requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
pydantic==2.12.5
python-dotenv==1.2.2
sentence-transformers==5.3.0
faiss-cpu==1.13.2
numpy==2.3.5
openai==2.30.0
tqdm==4.67.3
rank-bm25==0.2.2
slowapi==0.1.9

# Multi-modal document support
PyMuPDF==1.25.5
beautifulsoup4==4.13.4
python-docx==1.1.2

# Token-aware chunking (optional, for 'token' strategy)
tiktoken==0.9.0

# Observability & Monitoring
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-exporter-otlp==1.33.0
prometheus-client==0.22.1
structlog==25.4.0

# Semantic caching (optional)
redis>=5.0.0